What Flower app
What a flower is an app created using swift and UIKit and coreML. It identifies the flower that you've taken a photon and retrieves a short information from wikipedia.
Also used Cocoa Pods (Alamofire, SwiftyJSON) and the Wikipedia's open API.
The coreML file is not included because is a heavy file.